626: refactor names of class and methods
