+++ kafka-ui-e2e-checks/src/test/java/com/provectus/kafka/ui/SmokeTests.java
@@ -13,7 +13,7 @@ public class SmokeTests extends BaseTest {
     @Issue("380")
     void mainPageLoads() {
         pages.open()
-                .shouldBeOnPage();
+                .isOnPage();
         compareScreenshots("main");
     }
